tst_QHashFunctions: extend the consistency tests with pointer types
It's ... broken. Found and filed a bug. Added QEXPECT_FAIL() to
document the state of affairs, hopefully reminding us to fix these
things come Qt 7.

The multiple-inheritance thing is nothing we can do something about,
but keep it in as documentation.

Picking to 6.5 to verify this is not a regression since then.

Task-number: QTBUG-116986
Pick-to: 6.11 6.8 6.5
Change-Id: I44bc37d18f3f0c40be3dc269e5227a94f16c0e8d
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>