QIdentityProxyModel: store connection handles in a container
Port to PMF signal/slot syntax: compile-time type checking vs. run-time
string-based search. Given how many connections are made, this is both
safer and more readable.

Drop the _q_ prefix from method names, it was needed since
Q_PRIVATE_SLOT() was used for those methods.

Use QVLA since we know the size at compile time, (not a std::array like
other changes because some elements in the std::initializer_list are
going to be split out, so using a container with a push_back() is
simpler than fiddling with indexes to use operator[], see next commit in
this chain for details).

This is similar to c609b8dba0ae43aeba142164684536f808aad501, so a
similar performance improvement is expected.

Change-Id: Ia7e7d0b1f3c1a95bc399bd06de3b5c96e845b34d
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>