rhi: d3d11: Call Flush in endOffscreenFrame
If an application makes frequent calls to beginOffscreenFrame/
endOffscreenFrame without making any calls that implicitly flushes
queued-up commands (such as Present), we may spend too much memory
unless we have an explicit Flush. This happens because D3D11 defers
destruction of objects.

The issue appear with QtMultimedia's QVideoFrame::toImage, which
seems to leak memory on 6.5.2 and 6.4.3. It is not present on 6.6
because Flush() is called in endOffscreenFrame.

Fixes: QTBUG-111975
Change-Id: I3445836fac909ae9d46fb909d8ebc023a980a347
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>