Revise [since 6.7] to [since 6.5.3] for new CLDR enum items
Their introduction got picked back to 6.6 and 6.5 to fix a bug
reported as they were landing in dev; so 6.6 and dev are now being
updated to say [since 6.6] but they didn't make 6.5.[0-2], so they're
only [since 6.5.3] for the purposes of 6.5.

The lack of any pick-to here is, of course, deliberate.

Change-Id: I8ca1ca03bb3b9c4107fba3063946f35123b8f3ec
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>